requirements.txt
absl-py==2.0.0
altair==5.0.1
astunparse==1.6.3
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cycler==0.11.0
decorator==5.1.1
deepface==0.0.79
filelock==3.12.2
filetype==1.2.0
fire==0.5.0
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
gunicorn==21.2.0
h5py==3.8.0
idna==3.4
imageio==2.31.2
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
mdurl==0.1.2
mediapipe==0.9.0.1
mtcnn==0.1.1
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rich==13.6.0
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
six==1.16.0
smmap==5.0.1
soupsieve==2.4.1
streamlit==1.23.1
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tifffile==2021.11.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.66.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.6
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.15.0
zipp==3.15.0
